make some improvements
