--- src/entities/TodoContex/index.ts
@@ -1,1 +1,0 @@
-export * from './TodoContext';
